+++ server/blogDB.js
@@ -3,8 +3,8 @@
 var client = require('./client.js')('blog');
 var Vision = require('vision');
 var Handlebars = require('handlebars');
-// var Inert = require('inert');
 
+var handlers = require('./dashboardHandlers.js');
 
 exports.register = function(server, options, next) {
 
@@ -22,30 +22,8 @@ exports.register = function(server, options, next) {
         method: 'GET',
         path: '/dashboard/{user}',
         config: {
-            handler: function(request, reply) {
-                function retrieveBlogs(client, callback) {
-                    //uncomment following line and refresh page if you want to
-                    //flush DB
-                    // client.flushall();
-                    client.LRANGE('posts', 0, -1, function(err, reply) {
-                        if (err) {
-                            console.log(err);
-                        } else {
-
-                            callback(reply);
-                        }
-                    });
-                }
-                retrieveBlogs(client, function(postObjectsArray) {
-                    // typeof postObjectsArray is object here
-                    var parsedArray = postObjectsArray.map(function(el) {
-                        return JSON.parse(el);
-                    }).reverse();
-                    // var keysArray = Object.keys(postObjectsArray);
-                    reply.view('home', {
-                        data: parsedArray
-                    });
-                });
+            handler: function (request, reply){
+                handlers.getDashboard(request, reply);
             }
         }
     });
@@ -64,77 +42,18 @@ exports.register = function(server, options, next) {
         path: '/dashboard/{user}/{id}/delete',
         config: {
             handler: function(request, reply) {
-                console.log('reached handler');
-
-                function deletePost(client, objToBeDeleted, callback) {
-                    client.lrem('posts', 0, objToBeDeleted, function(err, reply) {
-                        if (err) {
-                            console.log(err);
-                        } else {
-                            callback(reply);
-                        }
-                    });
-                }
-                function retrieveBlogs(client, callback) {
-                    //uncomment following line and refresh page if you want to
-                    //flush DB
-                    // client.flushall();
-                    client.LRANGE('posts', 0, -1, function(err, reply) {
-                        if (err) {
-                            console.log(err);
-                        } else {
-
-                            callback(reply);
-                        }
-                    });
-                }
-                retrieveBlogs(client, function(postObjectsArray) {
-                    var id = request.params.id;
-                    var objToBeDeleted = postObjectsArray.filter(function(el) {
-                        return el.indexOf(id) > -1;
-                    })[0];
-                    deletePost(client, objToBeDeleted, function() {
-                        reply.redirect('/dashboard/' + request.params.user);
-                    });
-                });
+                handlers.deletePostHandler(request, reply);
             }
         }
     });
 
-    // adds date, author, id to db once publish button is clicked
-    //TODO check button compatibility
 
     server.route({
         method: 'POST',
         path: '/dashboard/{user}/publish',
         config: {
             handler: function(request, reply) {
-                let dateObj = new Date();
-                let day = dateObj.getDate();
-                let month = dateObj.getMonth() + 1;
-                let year = dateObj.getFullYear();
-                let today = day + '/' + month + '/' + year;
-                console.log(today);
-                request.payload.id = Date.now();
-                console.log('this posts id is: ' + request.payload.id);
-                request.payload.author = request.params.user;
-                request.payload.date = today;
-                var postObj = JSON.stringify(request.payload);
-
-                console.log('payload______>', request.payload, postObj);
-
-                function addToDB(client, callback) {
-                    client.rpush('posts', postObj, function(err, reply) {
-                        if (err) {
-                            console.log(err);
-                        } else {
-                            callback(reply);
-                        }
-                    });
-                }
-                addToDB(client, function() {
-                    reply.redirect('/dashboard/' + request.params.user);
-                });
+                handlers.publishPost(request, reply);
             }
         }
     });
@@ -143,17 +62,8 @@ exports.register = function(server, options, next) {
         method: 'GET',
         path: '/dashboard/{user}/edit-post/{id}',
         config: {
-            handler: function(request, reply) {
-                retrieveBlogs(client, function(blogPosts){
-                    // console.log(blogPosts);
-                    blogPosts.forEach(function(post){
-                        var postObj = JSON.parse(post);
-                        if(postObj.id.toString() === request.params.id){
-
-                            reply.view('new-post', {data: postObj});
-                        }
-                    });
-                });
+            handler: function(request, reply){
+                handlers.editPost(request, reply);
             }
         }
     });
@@ -162,36 +72,8 @@ exports.register = function(server, options, next) {
         method: 'POST',
         path: '/dashboard/{user}/save/{id}',
         config: {
-            handler: function(request, reply) {
-                retrieveBlogs(client, function(array){
-                    array.forEach(function(post){
-                        var postObj = JSON.parse(post);
-                        // console.log(request.params.id.toString(), ': is the id from the params should equal ' + postObj.id);
-                        if(request.params.id === postObj.id.toString()){
-                            client.lrem('posts', 0, JSON.stringify(postObj), function(err,response){
-                                console.log('removing from db');
-
-                                if(err) {
-                                    console.log(err);
-                                }
-                                else {
-                                    postObj.blog = request.payload.blog;
-                                    addToDB(client, JSON.stringify(postObj), function(){
-                                        console.log('adding to db');
-                                        reply.redirect('/dashboard/'+request.params.user);
-                                    });
-
-                                }
-                            });
-                        }
-                        // return JSON.stringify(postObj);
-                    });
-
-                    //TODO update the database with new array
-                    //     reply with correct screen
-
-                });
-                // reply.redirect('/dashboard/' + request.params.user);
+            handler: function(request, reply){
+                handlers.saveEditedPost(request, reply);
             }
         }
     });
@@ -199,25 +81,6 @@ exports.register = function(server, options, next) {
     next();
 };
 
-function retrieveBlogs(client, callback) {
-    client.LRANGE('posts', 0, -1, function(err, reply) {
-        if (err) {
-            console.log(err);
-        } else {
-            callback(reply);
-        }
-    });
-}
-
-function addToDB(client, postObj, callback) {
-    client.rpush('posts', postObj, function(err, reply) {
-        if (err) {
-            console.log(err);
-        } else {
-            callback(reply);
-        }
-    });
-}
 
 exports.register.attributes = {
     name: 'blogDB'
